backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
mangum==0.17.0
alembic==1.12.1
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
python-multipart==0.0.6
email-validator==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
redis==5.0.1
celery==5.3.4
bcrypt==3.2.2
